links abstraction of sidebar
